requirements.txt
absl-py==1.0.0
anaconda-navigator==2.1.1
appdirs==1.4.4
argh==0.26.2
astunparse==1.6.3
atomicwrites==1.4.0
backports.weakref==1.0.post1
bkcharts==0.2
black==19.10b0
boto==2.49.0
brotlipy==0.7.0
cachetools==4.2.4
certifi==2021.10.8
click==8.0.3
clyent==1.2.2
comtypes==1.1.10
conda==4.10.3
conda-build==3.21.6
conda-verify==3.4.2
cycler==0.10.0
cytoolz==0.11.0
daal4py==2021.3.0
dask==2021.10.0
distlib==0.3.4
entrypoints==0.3
et-xmlfile==1.1.0
Flask-Login==0.5.0
Flask-MySQLdb==0.2.0
flatbuffers==2.0
fonttools==4.25.0
gast==0.4.0
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.43.0
gunicorn==20.1.0
inflection==0.5.1
keras==2.7.0
Keras-Preprocessing==1.1.2
libclang==12.0.0
llvmlite==0.37.0
locket==0.2.1
Markdown==3.3.6
mccabe==0.6.1
mkl-fft==1.3.1
mkl-service==2.4.0
mpmath==1.2.1
munkres==1.1.4
mypy-extensions==0.4.3
mysqlclient==2.1.0
navigator-updater==0.2.1
nltk==3.6.5
oauthlib==3.1.1
pathspec==0.7.0
patsy==0.5.2
pep8==1.7.1
Pillow==8.4.0
pkginfo==1.7.1
platformdirs==2.4.1
ply==3.11
protobuf==3.19.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycosat==0.6.3
pycurl==7.44.1
pyls-spyder==0.4.0
pyodbc===4.0.0-unsupported
pytest==6.2.4
python-lsp-server==1.2.4
pytz==2021.3
pywin32==228
PyYAML==6.0
requests-oauthlib==1.3.0
rsa==4.8
scikit-image==0.18.3
simplegeneric==0.8.1
sip==4.19.13
Sphinx==4.2.0
tables==3.6.1
TBB==0.2
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.7.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.23.1
termcolor==1.1.0
terminado==0.9.4
urllib3==1.26.7
virtualenv==20.13.0
webencodings==0.5.1
win-unicode-console==0.5
wincertstore==0.2
xlwings==0.24.9
xlwt==1.3.0
zict==2.0.0
zope.event==4.5.0
